requirements.txt
Flask==2.2.3
Jinja2==3.1.2
pandas==1.5.3
scikit-surprise==1.1.3
